+++ babel.config.js
@@ -1,5 +1,3 @@
 module.exports = {
   presets: ['module:@react-native/babel-preset'],
-  // Reanimated's plugin must be listed last.
-  plugins: ['react-native-reanimated/plugin'],
 };